+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### v3.0.194 (Jul 29, 2020)
+* Added `is_muted` property in `SBDMember` class
+* Added `applicationAttributes` property in `SBDAppInfo` class
+* Improved stability
+
 ### v3.0.193 (Jul 22, 2020)
 * Added open graph feature along with `SBDOGMetaData` and `SBDOGImage` class
 * Added `ogMetaData` property in `SBDBaseMessage`